# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloud Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CloudService {
  id: number;
  provider: string;
  name: string;
  cpu: number;
  ramGb: number;
  storageGb: number;
  storageType: string;
  bandwidthTb: number;
  priceMonthly: number;
  region: string;
  isActive: boolean;
  createdAt: string;
}

export type ServerOrderStatus =
  (typeof ServerOrderStatus)[keyof typeof ServerOrderStatus];

export const ServerOrderStatus = {
  Pending: "Pending",
  Active: "Active",
  Failed: "Failed",
} as const;

export interface ServerOrder {
  id: number;
  userId: string;
  cloudServiceId: number;
  status: ServerOrderStatus;
  requestedRegion: string;
  notes?: string | null;
  externalOrderId?: string | null;
  providerResponse?: string | null;
  createdAt: string;
  cloudService?: CloudService | null;
}

export interface CreateOrderBody {
  cloudServiceId: number;
  requestedRegion: string;
  notes?: string | null;
}

export interface DashboardStats {
  totalOrders: number;
  activeOrders: number;
  pendingOrders: number;
  failedOrders: number;
  totalServices: number;
  recentOrders: ServerOrder[];
}

export interface ProviderStat {
  provider: string;
  count: number;
  avgPrice: number;
}

export type ListServicesParams = {
  provider?: string;
  region?: string;
  minPrice?: number;
  maxPrice?: number;
};