# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloud Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all cloud services
 */
export const ListServicesQueryParams = zod.object({
  provider: zod.coerce.string().optional(),
  region: zod.coerce.string().optional(),
  minPrice: zod.coerce.number().optional(),
  maxPrice: zod.coerce.number().optional(),
});

export const ListServicesResponseItem = zod.object({
  id: zod.number(),
  provider: zod.string(),
  name: zod.string(),
  cpu: zod.number(),
  ramGb: zod.number(),
  storageGb: zod.number(),
  storageType: zod.string(),
  bandwidthTb: zod.number(),
  priceMonthly: zod.number(),
  region: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListServicesResponse = zod.array(ListServicesResponseItem);

/**
 * @summary Get a cloud service by ID
 */
export const GetServiceParams = zod.object({
  id: zod.coerce.number(),
});

export const GetServiceResponse = zod.object({
  id: zod.number(),
  provider: zod.string(),
  name: zod.string(),
  cpu: zod.number(),
  ramGb: zod.number(),
  storageGb: zod.number(),
  storageType: zod.string(),
  bandwidthTb: zod.number(),
  priceMonthly: zod.number(),
  region: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get current user's orders
 */
export const ListMyOrdersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  cloudServiceId: zod.number(),
  status: zod.enum(["Pending", "Active", "Failed"]),
  requestedRegion: zod.string(),
  notes: zod.string().nullish(),
  externalOrderId: zod.string().nullish(),
  providerResponse: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  cloudService: zod
    .object({
      id: zod.number(),
      provider: zod.string(),
      name: zod.string(),
      cpu: zod.number(),
      ramGb: zod.number(),
      storageGb: zod.number(),
      storageType: zod.string(),
      bandwidthTb: zod.number(),
      priceMonthly: zod.number(),
      region: zod.string(),
      isActive: zod.boolean(),
      createdAt: zod.coerce.date(),
    })
    .nullish(),
});
export const ListMyOrdersResponse = zod.array(ListMyOrdersResponseItem);

/**
 * @summary Request a server
 */
export const CreateOrderBody = zod.object({
  cloudServiceId: zod.number(),
  requestedRegion: zod.string(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get order by ID
 */
export const GetOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  cloudServiceId: zod.number(),
  status: zod.enum(["Pending", "Active", "Failed"]),
  requestedRegion: zod.string(),
  notes: zod.string().nullish(),
  externalOrderId: zod.string().nullish(),
  providerResponse: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  cloudService: zod
    .object({
      id: zod.number(),
      provider: zod.string(),
      name: zod.string(),
      cpu: zod.number(),
      ramGb: zod.number(),
      storageGb: zod.number(),
      storageType: zod.string(),
      bandwidthTb: zod.number(),
      priceMonthly: zod.number(),
      region: zod.string(),
      isActive: zod.boolean(),
      createdAt: zod.coerce.date(),
    })
    .nullish(),
});

/**
 * @summary Get dashboard statistics for the current user
 */
export const GetDashboardStatsResponse = zod.object({
  totalOrders: zod.number(),
  activeOrders: zod.number(),
  pendingOrders: zod.number(),
  failedOrders: zod.number(),
  totalServices: zod.number(),
  recentOrders: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      cloudServiceId: zod.number(),
      status: zod.enum(["Pending", "Active", "Failed"]),
      requestedRegion: zod.string(),
      notes: zod.string().nullish(),
      externalOrderId: zod.string().nullish(),
      providerResponse: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      cloudService: zod
        .object({
          id: zod.number(),
          provider: zod.string(),
          name: zod.string(),
          cpu: zod.number(),
          ramGb: zod.number(),
          storageGb: zod.number(),
          storageType: zod.string(),
          bandwidthTb: zod.number(),
          priceMonthly: zod.number(),
          region: zod.string(),
          isActive: zod.boolean(),
          createdAt: zod.coerce.date(),
        })
        .nullish(),
    }),
  ),
});

/**
 * @summary Get stats grouped by cloud provider
 */
export const GetProviderStatsResponseItem = zod.object({
  provider: zod.string(),
  count: zod.number(),
  avgPrice: zod.number(),
});
export const GetProviderStatsResponse = zod.array(GetProviderStatsResponseItem);