# lib/api-zod/src/generated/types/cloudService.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloud Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CloudService {
  id: number;
  provider: string;
  name: string;
  cpu: number;
  ramGb: number;
  storageGb: number;
  storageType: string;
  bandwidthTb: number;
  priceMonthly: number;
  region: string;
  isActive: boolean;
  createdAt: Date;
}