# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloud Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ServerOrder } from "./serverOrder";

export interface DashboardStats {
  totalOrders: number;
  activeOrders: number;
  pendingOrders: number;
  failedOrders: number;
  totalServices: number;
  recentOrders: ServerOrder[];
}