# lib/api-zod/src/generated/types/serverOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloud Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CloudService } from "./cloudService";
import type { ServerOrderStatus } from "./serverOrderStatus";

export interface ServerOrder {
  id: number;
  userId: string;
  cloudServiceId: number;
  status: ServerOrderStatus;
  requestedRegion: string;
  notes?: string | null;
  externalOrderId?: string | null;
  providerResponse?: string | null;
  createdAt: Date;
  cloudService?: CloudService | null;
}